added geohack winner to demo site
