+++ app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@bayer/ol-kit",
+  "name": "ol-kit-demo",
   "version": "1.0.1",
   "license": "BSD",
   "description": "Mapping components & utils built with openlayers + react",
@@ -42,10 +42,8 @@
   "bugs": "https://github.com/MonsantoCo/ol-kit/issues",
   "main": "./core/index.js",
   "scripts": {
-    "build": "rm -rf core/ && babel src/ -d core",
-    "dev": "rm -rf core/ && { babel --watch src/ -d core; } | tee >&2",
-    "app:build": "webpack --config app/webpack.config.js",
-    "app": "webpack-dev-server --hot --inline --config app/webpack.config.js",
+    "app:build": "webpack --config webpack.config.js",
+    "app": "NODE_ENV=development webpack-dev-server --hot --inline --config webpack.config.js",
     "docs": "rm -rf docs/ && jsdoc -r -c jsdoc.json && cp src-docs/overrides/* docs/",
     "integrations": "./app/integration.sh",
     "lint": "eslint src/**/*.js",
@@ -61,94 +59,36 @@
     "url": "git@github.com:MonsantoCo/ol-kit.git"
   },
   "dependencies": {
-    "@date-io/moment": "~1.3.13",
+    "@bayer/ol-kit": "^1.0.1",
     "@material-ui/core": "~4.9.12",
     "@material-ui/icons": "~4.9.1",
     "@material-ui/pickers": "~3.2.10",
     "@material-ui/styles": "~4.9.10",
-    "@mdi/js": "~5.1.45",
-    "@mdi/react": "1.4.0",
-    "@turf/bbox": "~6.0.1",
-    "@turf/centroid": "~6.0.2",
-    "@turf/helpers": "~6.1.4",
-    "clipboard": "^2.0.6",
-    "geostyler-openlayers-parser": "^2.1.0",
-    "lodash.clonedeep": "~4.5.0",
-    "lodash.debounce": "~4.0.8",
-    "lodash.escaperegexp": "~4.1.2",
-    "lodash.isequal": "~4.5.0",
-    "moment": "~2.26.0",
-    "nanoid": "~2.1.11",
+    "file-saver": "^2.0.2",
+    "formik": "^2.1.6",
+    "mui-datatables": "^3.4.1",
     "ol": "^6.4.3",
-    "prop-types": "~15.7.2",
-    "qs": "~6.9.3",
-    "react-color": "~2.18.0",
-    "react-draggable": "^4.4.3",
-    "react-hook-form": "~5.7.2",
-    "react-select": "~3.1.0",
-    "shp-write": "~0.3.2",
-    "ugh": "^0.1.1",
-    "uuid": "~7.0.3"
+    "react-bootstrap": "^1.3.0",
+    "recharts": "^1.8.5",
+    "ugh": "^0.1.1"
   },
   "devDependencies": {
-    "@babel/cli": "~7.4.3",
-    "@babel/core": "~7.7.7",
-    "@babel/plugin-proposal-class-properties": "~7.7.4",
-    "@babel/plugin-proposal-optional-chaining": "~7.9.0",
-    "@babel/plugin-transform-runtime": "~7.6.2",
-    "@babel/polyfill": "~7.8.3",
-    "@babel/preset-env": "~7.7.1",
-    "@babel/preset-react": "~7.7.0",
-    "@babel/runtime": "~7.7.2",
-    "@hot-loader/react-dom": "~16.13.0",
-    "@testing-library/dom": "~7.0.4",
-    "@testing-library/jest-dom": "~5.5.0",
-    "@testing-library/react": "~10.0.1",
     "apollo-boost": "^0.4.9",
-    "babel-eslint": "~10.0.1",
-    "babel-jest": "~24.9.0",
-    "babel-loader": "~8.0.6",
-    "babel-plugin-inline-react-svg": "~1.1.0",
-    "babel-plugin-module-resolver": "~3.2.0",
-    "better-docs": "~1.4.7",
-    "child_process": "~1.0.2",
     "css-loader": "~3.6.0",
-    "enzyme": "~3.10.0",
-    "enzyme-adapter-react-16": "~1.15.2",
-    "enzyme-to-json": "~3.4.4",
     "eslint": "~6.8.0",
-    "eslint-config-standard": "~14.1.1",
-    "eslint-loader": "~4.0.2",
-    "eslint-plugin-import": "~2.20.2",
-    "eslint-plugin-node": "~11.0.0",
-    "eslint-plugin-promise": "~3.5.0",
-    "eslint-plugin-react": "~7.19.0",
-    "eslint-plugin-react-hooks": "^4.1.0",
-    "eslint-plugin-standard": "~4.0.1",
     "file-loader": "^6.1.0",
     "graphql": "^14.7.0",
     "html-webpack-plugin": "~4.3.0",
-    "husky": "~4.2.5",
     "jest": "~25.4.0",
-    "jest-canvas-mock": "~2.2.0",
-    "jest-environment-jsdom": "~26.0.1",
-    "jest-environment-jsdom-global": "~2.0.2",
-    "jest-fetch-mock": "~3.0.3",
-    "jest-puppeteer": "~4.4.0",
-    "jest-styled-components": "~7.0.2",
     "jsdoc": "3.6.3",
-    "parcel": "2.0.0-alpha.3.2",
     "puppeteer": "^4.0.1",
     "react": "~16.11.0",
-    "react-apollo": "^3.1.5",
     "react-dom": "~16.11.0",
     "react-hot-loader": "~4.12.21",
     "react-router-dom": "^5.2.0",
     "runjs": "~4.4.2",
-    "semver": "~7.3.2",
     "styled-components": "~5.0.1",
     "webpack": "~4.43.0",
-    "webpack-cli": "~3.3.12",
     "webpack-dev-server": "~3.11.0",
     "yup": "^0.29.3"
   },
